redteam_agent/requirements.txt
langgraph>=0.3.20
langchain-core>=0.3.60
langchain-openai>=0.3.16
pydantic>=2.9.0
typer>=0.12.3
rich>=13.9.0
networkx>=3.3
chromadb>=0.5.5
numpy>=1.26.0
docker>=7.1.0
sqlalchemy>=2.0.35
httpx>=0.27.2
tenacity>=9.0.0
pyyaml>=6.0.2
orjson>=3.10.7
aiosqlite>=0.20.0
pytest>=8.3.3
msgpack>=1.1.0
